Make clippy happy when no_std is used
