+++ roaring/src/bitmap/store/bitmap_store.rs
@@ -356,6 +356,7 @@ impl BitmapStore {
         BitmapIter::new(self.bits)
     }
 
+    #[cfg(feature = "std")]
     pub fn as_array(&self) -> &[u64; BITMAP_LENGTH] {
         &self.bits
     }
