+++ roaring/src/bitmap/store/mod.rs
@@ -38,6 +38,7 @@ impl Store {
         Store::Array(ArrayStore::new())
     }
 
+    #[cfg(feature = "std")]
     pub fn with_capacity(capacity: usize) -> Store {
         if capacity <= ARRAY_LIMIT as usize {
             Store::Array(ArrayStore::with_capacity(capacity))
